Move MockMemoryReader to separate file (phpcs one-class-per-file)
