+++ tests/Lib/Process/MemoryReader/BufferedMemoryReaderTest.php
@@ -245,31 +245,3 @@ class BufferedMemoryReaderTest extends TestCase
         $this->assertSame('NEW!', FFI::string($result, 4));
     }
 }
-
-/**
- * @internal test helper
- */
-class MockMemoryReader implements MemoryReaderInterface
-{
-    /** @var array<string, string> */
-    public array $data = [];
-    public int $readCount = 0;
-
-    public function preload(int $address, string $bytes): void
-    {
-        $this->data[$address . ':' . strlen($bytes)] = $bytes;
-    }
-
-    #[\Override]
-    public function read(int $pid, int $remote_address, int $size): FFI\CData
-    {
-        $this->readCount++;
-        $key = $remote_address . ':' . $size;
-        $bytes = $this->data[$key] ?? str_repeat("\0", $size);
-        $buf = FFI::new("char[$size]");
-        assert($buf !== null);
-        FFI::memcpy($buf, $bytes, $size);
-        /** @var \FFI\CArray<int> */
-        return $buf;
-    }
-}
